requirements.txt
# ── Core data pipeline ────────────────────────────────────────────────────────
alpaca-py==0.43.2
pandas==3.0.1
numpy==2.4.2
pyarrow==23.0.1
python-dotenv==1.2.1
requests==2.32.5

# ── Market data sources ───────────────────────────────────────────────────────
yfinance==1.2.0
fredapi==0.5.2
Nasdaq-Data-Link==1.0.4

# ── HTML / XML parsing (SEC EDGAR) ────────────────────────────────────────────
lxml==6.0.2
beautifulsoup4==4.14.3

# ── Telegram bot ──────────────────────────────────────────────────────────────
python-telegram-bot==22.6

# ── Charts (reports) ──────────────────────────────────────────────────────────
matplotlib==3.10.8
